+++ src/main/java/de/stw/core/clock/Clock.java
@@ -5,5 +5,5 @@ import java.util.concurrent.TimeUnit;
 public interface Clock {
     Tick nextTick();
     Tick getCurrentTick();
-    Tick getTick(int duration, TimeUnit timeUnit);
+    Tick getTick(long duration, TimeUnit timeUnit);
 }
